Moved Lti2 to Lti\v2
