+++ LtiLibrary.AspNetCore.Tests/Profiles/ToolConsumerProfileController.cs
@@ -3,7 +3,6 @@ using System.Threading.Tasks;
 using LtiLibrary.AspNetCore.Profiles;
 using LtiLibrary.NetCore.Common;
 using LtiLibrary.NetCore.Lti.v2;
-using LtiLibrary.NetCore.Lti2;
 using NodaTime;
 
 namespace LtiLibrary.AspNetCore.Tests.Profiles
