+++ LtiLibrary.NetCore.Tests/ToolConsumerProfileTests.cs
@@ -2,7 +2,6 @@
 using LtiLibrary.NetCore.Tests.SimpleHelpers;
 using LtiLibrary.NetCore.Common;
 using LtiLibrary.NetCore.Lti.v2;
-using LtiLibrary.NetCore.Lti2;
 using NodaTime;
 using Xunit;
