+++ LtiLibrary.NetCore/Lti/v2/Contact.cs
@@ -1,6 +1,6 @@
 ﻿using Newtonsoft.Json;
 
-namespace LtiLibrary.NetCore.Lti2
+namespace LtiLibrary.NetCore.Lti.v2
 {
     /// <summary>
     /// Contact information for the associated object.
