+++ LtiLibrary.NetCore/Lti/v2/ServiceOwner.cs
@@ -2,7 +2,7 @@
 using LtiLibrary.NetCore.Common;
 using Newtonsoft.Json;
 
-namespace LtiLibrary.NetCore.Lti2
+namespace LtiLibrary.NetCore.Lti.v2
 {
     /// <summary>
     /// Represents an IMS ServiceOwner object.
